README.md
This GitHub profile README introduces Raksith Sivakumar, a Computer Science student at Bannari Amman Institute of Technology with interests in Cyber Security and Full Stack Development. The profile features a dynamic layout, including an animated image, badges linking to LinkedIn and Telegram, a list of coding languages and tools used, and a display of GitHub stats and activity graph.